Moved Activity model to appropriate namespace.
