+++ Application/Activities/List.cs
@@ -1,4 +1,5 @@
 ﻿using Domain;
+using Domain.Entities;
 using FluentResults;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
