+++ Application/Core/MappingProfiles.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Domain;
+using Domain.Entities;
 
 namespace Application.Core;
